+++ src/pokemonoceanblue/StatusEffect.java
@@ -1,13 +1,13 @@
 package pokemonoceanblue;
 
 public class StatusEffect {
-    public static int UNAFFLICTED = 0;
-    public static int PARALYSIS = 1;
-    public static int SLEEP = 2;
-    public static int FROZEN = 3;
-    public static int BURN = 4;
-    public static int POISON = 5;
-    public static int BADLY_POISON = 6;
-    public static int CURSE = 7;
-    public static int CONFUSION = 8;
+    public static byte UNAFFLICTED = 0;
+    public static byte PARALYSIS = 1;
+    public static byte SLEEP = 2;
+    public static byte FROZEN = 3;
+    public static byte BURN = 4;
+    public static byte POISON = 5;
+    public static byte BADLY_POISON = 6;
+    public static byte CURSE = 7;
+    public static byte CONFUSION = 8;
 }
